% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getStations.R
\name{getStations}
\alias{getStations}
\title{Returns the Personal Weather Stations nearby input provided}
\usage{
getStations(zip, radius = 10)
}
\arguments{
\item{Zip}{Code and the radius in miles}
}
\value{
data frame containing weather stations data
}
\description{
This function accepts the zip code and a radius in miles and
             returns all the weather stations in that range in a table.
}
\author{
Maruthi Ram Nadakuduru
}

