% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeStationSubtable.R
\name{makeStationSubtable}
\alias{makeStationSubtable}
\title{Create subtable of PWS}
\usage{
makeStationSubtable(pwStations, newRadius = NA, numberToKeep = NA,
  nearest = TRUE, stationNames = NA)
}
\arguments{
\item{pwStations}{A PWStations object with the data to be filtered.}

\item{newRadius}{A positive numeric value representing the max distance_mi.
Anything greater or equal to the original radius will result in no
changes.}

\item{numberToKeep}{An integer value specifying the number of stations to
keep, sorted by distance.}

\item{nearest}{Boolean value indicating if stations should be kept
according to the nearest(sorted by increasing distance_mi) or
farthest (sorted by decreasing distance_mi). Default is TRUE.}

\item{stationNames}{A vector of the names/IDs of specific stations in the
table.}
}
\value{
A PWStations object representing the result of applying the
        specified criteria to the original object.
}
\description{
Returns a PWStations object with the specified filter applied
             to the original PWStations. If the original object contained
             weather data, this will also be filtered to the appropriate
             stations.
Filter options are one of the following:
\itemize{
\item Reducing the radius of the original query, so that only stations
\item within the new radius are kept.
\item Choosing a number of closest or farthest stations from the original
\item query location.
\item Choosing a set of stations by name.
}
}
\examples{
data(charlotteStations)
# Decrease radius to 5 miles
newRadTable <- makeStationSubtable(charlotteStations,
                                   newRadius = 5)
newRadTable$stations

# Keep 3 closest
closestTable <- makeStationSubtable(charlotteStations,
                                    numberToKeep = 3)
closestTable$stations

# Keep 3 farthest
farthestTable <- makeStationSubtable(charlotteStations,
                                     numberToKeep = 3,
                                     nearest = FALSE)
farthestTable$stations

# Keep specific stations
specificTable <- makeStationSubtable(charlotteStations,
                                     stationNames = c("KNCCHARL71",
                                                      "KNCCHARL83"))
specificTable$stations
}
\author{
Jared Casale
}
\seealso{
\code{\link{getStations}} for a method that will generate an object.
         See \code{\link{loadWeatherData}} to load weather data for the
         selected PWS stations.
}

