% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ProjectPWS.R
\docType{package}
\name{ProjectPWS}
\alias{ProjectPWS}
\alias{ProjectPWS-package}
\title{ProjectPWS: A package for extracting weather data from
personal weather stations (PWS) from the Weather
Underground API.}
\description{
ProjectPWS: A package for extracting weather data from
personal weather stations (PWS) from the Weather
Underground API.
}

