% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PWStations-class.R
\docType{class}
\name{PWStations-class}
\alias{PWStations-class}
\title{Function defining the PWStations S4 class}
\description{
This function includes all the definitions of the PWStations
             S4 class.
}
\author{
Maruthi Ram Nadakuduru
}

