% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PWStations-class.R
\docType{data}
\name{PWStations}
\alias{PWStations}
\title{R6 Class used to hold the result of a Wunderground PWS query.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x00000000089f1f68> 
 - attr(*, "name")= chr "PWStations_generator"
}}
\usage{
PWStations
}
\description{
This class holds the result of a WUnderground API query. The parameters imply
the query used (i.e. if zip is set, a zip lookup was performed) and the resulting stations
are stored in a stations data table.
}
\examples{
\dontrun{
atLongStations <- getStations(latlong = c(35.229, 80.8433), radius = 2) # Charlotte
}
}
\author{
Maruthi Ram Nadakuduru, Jared Casale
}
\keyword{datasets}

