% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readPWSTable.R
\name{readPWSTable}
\alias{readPWSTable}
\title{Load PWS Table}
\usage{
readPWSTable(fileName)
}
\arguments{
\item{fileName}{Name of file where saved PWStation object exists.}
}
\description{
This function reads a saved PWS table created by user from
             the specified location.
}
\author{
Maruthi Ram Nadakuduru
}

