% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/savePWSTable.R
\name{savePWSTable}
\alias{savePWSTable}
\title{Save PWS Table}
\usage{
savePWSTable(pwsTable, fileName)
}
\description{
This function saves the PWS table created by user in the specified location
}
\author{
Maruthi Ram Nadakuduru
}

