% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/startShiny.R
\name{startShiny}
\alias{startShiny}
\title{This is a test hook to start the shiny app}
\usage{
startShiny()
}
\description{
This is a test hook to start the shiny app
}
\author{
Maruthi Ram Nadakuduru, Jared Casale
}

