% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getStations.R
\name{getStations}
\alias{getStations}
\title{Returns the Personal Weather Stations nearby input provided}
\usage{
getStations(zip, radius = 10, state, country)
}
\arguments{
\item{Zip}{Code and the radius in kilometers}
}
\value{
PWStations class with weather stations data
}
\description{
This function accepts the zip code and a radius in km and
             returns all the stations in that range in the PWStations class.
}
\author{
Maruthi Ram Nadakuduru, Jared Casale
}

