% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadWeatherData.R
\name{loadWeatherData}
\alias{loadWeatherData}
\title{Load PWS Weather Data}
\usage{
loadWeatherData(pwStations, startDate, endDate = NA, startHour = NA,
  endHour = NA, weatherVars = c("tempi", "hum", "wspdi", "pressure"),
  stationLimit = -1L)
}
\arguments{
\item{pwStations}{PWStation object containing the stations that require data.}

\item{startDate}{Beginning of time period of interest. Must be a string with
format mm/dd/YYYY.}

\item{endDate}{End of time period of interest. If specified, must be a
string with format mm/dd/YYYY.}

\item{startHour}{First hour of period of interest.
Must be numeric between 0 and 23.}

\item{endHour}{Last hour of period of interest.
Must be numeric between 0 and 23.}

\item{weatherVars}{Variables to retrieve, must be one of "tempi"
(temperature imperial i.e. Fahrenheit), "hum" (humidity),
"wspid" (Weather speed imperial), "pressure". Plain-text
weather conditions ("conds") will also be extracted.}

\item{stationLimit}{This restricts to the first stationLimit stations.
-1 signals to retrieve all in the table.}
}
\value{
List of data frames, one for each weather station in the original
        PWStations object. The name of each element is the station's ID.
Note that this is also accessible from the supplied PWStations object.
}
\description{
This function will query the wunderground site and retrieve the
             weather data for each PW Station, over the specified time
             period.
}
\examples{
\dontrun{
stations <- getStations(zip = "98107",
                        radius = 2)
loadWeatherData(stations,
                startDate = "3/3/2015",
                startHour = 9,
                endHour = 5,
                stationLimit = 1)

length(stations$weatherData)
nrow(stations$weatherData[[1]])

loadWeatherData(stations,
                startDate = "3/1/2015",
                endDate = "3/3/2015",
                stationLimit = 3)

length(stations$weatherData)
nrow(stations$weatherData[[1]])
}
}
\author{
Jared Casale
}

