% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PWStations-class.R
\docType{data}
\name{PWStations}
\alias{PWStations}
\title{Function defining the PWStations R6 class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0000000005db7420> 
 - attr(*, "name")= chr "PWStations_generator"
}}
\usage{
PWStations
}
\description{
This function includes all the definitions of the PWStations
             R6 class.
}
\author{
Maruthi Ram Nadakuduru, Jared Casale
}
\keyword{datasets}

