% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validateTrimAndFill.R
\name{validateTrimAndFill}
\alias{validateTrimAndFill}
\title{Clean and fill PWS weather data}
\usage{
validateTrimAndFill(pwStations, stopAfterValidation = FALSE,
  stopAfterTrim = FALSE)
}
\arguments{
\item{pwStations}{PWStations object with loaded weather data.}

\item{stopAfterValidation}{Will stop after replacing values outside the
reasonable range with NA and return the resulting list of tables.}

\item{stopAfterTrim}{Will stop after trimming columns with less than 2 data
points.}
}
\value{
A list of data tables representing the original weather data with
         the specified operations applied.
}
\description{
Takes a PWStation object and loops through all of its
             stored weather data. For each weather data table, it will:
\itemize{
\item  Check all values and determine if they are within a (generously)
       reasonable range. If they are not, they will be replaced with an NA.
\item  Trim each table so that any columns with less than 2 data points
       are removed.
\item  Fill in missing values by taking the mean of 5 imputations using
       the \code{\link[Amelia]{amelia}} function.
}
}
\examples{
data(charlotteStations)

# Just keep two stations
lessStations <- makeStationSubtable(charlotteStations,
                       stationNames = c("KNCCHARL71", "KNCCHARL83"))

# Just validate
justVal <- validateTrimAndFill(charlotteStations, stopAfterValidation = TRUE)
sum(is.na(justVal[[1]]))

# Just validate and trim
justTrim <- validateTrimAndFill(charlotteStations, stopAfterTrim = TRUE)
sum(is.na(justTrim[[1]]))

# Validate, trim and then fill missing data
filled <- validateTrimAndFill(charlotteStations)
sum(is.na(filled[[1]]))
}
\author{
Jared Casale
}

